Move tinyurl helpers into core module
